Modificaciones

+++ MantenimientoWeb/MantenimientoWeb-ejb/src/main/java/sv/edu/uesocc/ingenieria/tpi2018/entities/OrdenTrabajoPorTrabajadoresPK.java
@@ -54,14 +54,13 @@ public class OrdenTrabajoPorTrabajadoresPK implements Serializable {
     @Override
     public int hashCode() {
         int hash = 0;
-        hash += (int) idOrdenTrabajo;
-        hash += (int) idTrabajadores;
+        hash += idOrdenTrabajo;
+        hash += idTrabajadores;
         return hash;
     }
 
     @Override
     public boolean equals(Object object) {
-        // TODO: Warning - this method won't work in the case the id fields are not set
         if (!(object instanceof OrdenTrabajoPorTrabajadoresPK)) {
             return false;
         }
